+++ src/Chirp.Razor/scripts/initDB.sh
@@ -1,3 +1,3 @@
 #!/usr/bin/env bash
-sqlite3 ../data/cheepDatabase.db < data/schema.sql
-sqlite3 ../data/cheepDatabase.db < data/dump.sql
+sqlite3 /tmp/cheepDatabase.db < data/schema.sql
+sqlite3 /tmp/cheepDatabase.db < data/dump.sql
